Move query client initialization out of routes.tsx into queryClient.ts (#3986)
queryClient.ts now self-initializes via a top-level await on the user's
setupFn, then synchronously creates the QueryClient. routes.tsx is no
longer responsible for orchestrating unrelated setup.

The configureQueryClient state lives in a new leaf module
queryClient.config.ts so the cyclic import between queryClient.ts and the
user's setupFn module resolves safely (no TDZ on the let state).
