+++ waspc/data/Generator/templates/sdk/wasp/client/operations/hooks.ts
@@ -9,7 +9,7 @@ import {
 } from "@tanstack/react-query";
 import { Action, Query } from "./rpc";
 import { makeQueryCacheKey } from "./queries/core";
-export { configureQueryClient } from "./queryClient";
+export { configureQueryClient } from "./queryClient.config";
 
 // PUBLIC API
 export function useQuery<Input, Output>(
