+++ waspc/data/Generator/templates/sdk/wasp/client/operations/index.ts
@@ -16,7 +16,5 @@ export {
     // PUBLIC API
     configureQueryClient,
     // PRIVATE API (framework code)
-    initializeQueryClient,
-    // PRIVATE API (framework code)
     queryClientInitialized
 } from './queryClient'
