+++ waspc/data/Generator/templates/sdk/wasp/client/operations/queryClient.ts
@@ -1,34 +1,21 @@
-import { QueryClient, QueryClientConfig } from '@tanstack/react-query'
+{{={= =}=}}
+import { QueryClient } from '@tanstack/react-query'
+import { freezeAndGetConfig } from './queryClient.config'
+
+{=# setupFn.isDefined =}
+{=& setupFn.importStatement =}
+
+await {= setupFn.importIdentifier =}()
+{=/ setupFn.isDefined =}
 
 const defaultQueryClientConfig = {};
 
-let queryClientConfig: QueryClientConfig,
-  resolveQueryClientInitialized: (...args: any[]) => any,
-  isQueryClientInitialized: boolean;
-
-// PRIVATE API (framework code)
-export const queryClientInitialized: Promise<QueryClient> = new Promise(
-  (resolve) => {
-    resolveQueryClientInitialized = resolve;
-  }
+const queryClient = new QueryClient(
+  freezeAndGetConfig() ?? defaultQueryClientConfig
 );
 
-// PUBLIC API
-export function configureQueryClient(config: QueryClientConfig): void {
-  if (isQueryClientInitialized) {
-    throw new Error(
-      "Attempted to configure the QueryClient after initialization"
-    );
-  }
-
-  queryClientConfig = config;
-}
-
 // PRIVATE API (framework code)
-export function initializeQueryClient(): void {
-  const queryClient = new QueryClient(
-    queryClientConfig ?? defaultQueryClientConfig
-  );
-  isQueryClientInitialized = true;
-  resolveQueryClientInitialized(queryClient);
-}
+export const queryClientInitialized: Promise<QueryClient> = Promise.resolve(queryClient);
+
+// PUBLIC API
+export { configureQueryClient } from './queryClient.config';
